Use sbt to generate Eclipse project files.
Add sbteclipse to your plugin definition file at `~/.sbt/0.13/plugins/somefilename.sbt` by inserting this:
```
addSbtPlugin("com.typesafe.sbteclipse" % "sbteclipse-plugin" % "2.4.0")
```
This adds the `eclipse` command to the sbt shell. In Eclipse, use the "import
existing project into workspace" function. Do not copy the project to the
workspace.

This commit adds an Eclipse-specific config item to `build.sbt`, namely to
generate a Java project not a Scala project as is the default. Having the
sbteclipse plugin globally enabled, not project specific, seems nonsensical, but
is recommended. I expect the sbt people know something I don't, so let's just go
with it.